pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-custom"
version = "0.0.1"
description = "Some personal cross-project code to implement minimal agentic loops, tooling, and one-shot LLM usage in a fast and familiar way."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "anthropic==0.84.0",
    "ollama==0.6.1",
    "openai==2.14.0",
    "pydantic==2.12.5",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Optional dependencies can be added here in the future

[tool.setuptools]
# Automatically discover all packages and subpackages
packages = { find = {} }

[tool.setuptools.package-dir]
"" = "."

# Exclude playgrounds directory from package distribution
[tool.setuptools.exclude-package-data]
"*" = ["__pycache__/*"]
"playgrounds" = ["*"]

